requirements.txt
pandas==2.2.2
scikit-learn=1.5.0
mysql-connector-python==8.4.0
streamlit==1.36.0